Fixing routes navigations
