+++ src/app/(auth)/_layout.tsx
@@ -1,0 +1,14 @@
+import { Stack } from "expo-router";
+
+export default function AuthLayout() {
+    return (
+        <Stack
+            screenOptions={{
+                headerShown: false, // Hide the default header
+            }}
+        >
+            <Stack.Screen name="/login" />
+            <Stack.Screen name="/signup" />
+        </Stack>
+    );
+}
