+++ src/app/_layout.tsx
@@ -8,8 +8,6 @@ export default function RootLayout() {
     <QueryClientProvider client={queryClient}>
       <Stack screenOptions={{ headerShown: false }}>
         <Stack.Screen name="/" />
-        <Stack.Screen name="auth/login" />
-        <Stack.Screen name="auth/signup" />
       </Stack>
     </QueryClientProvider>
   );
